requirements.txt
# Jarvis AI Desktop Assistant — Python dependencies
# Install with:  pip install -r requirements.txt

# GUI framework
PyQt5>=5.15.0

# Speech recognition (wake word + command capture)
SpeechRecognition>=3.10.0

# Text-to-speech (online — used by your speak() function)
gTTS>=2.5.0

# Audio playback for gTTS output
pygame>=2.5.0

# Offline TTS fallback (already in your main.py)
pyttsx3>=2.90

# AI / LLM (Gemini)
google-genai>=1.0.0

# News API HTTP client
requests>=2.31.0

# PyAudio — microphone access for SpeechRecognition
# On Linux:   sudo apt install portaudio19-dev && pip install pyaudio
# On macOS:   brew install portaudio && pip install pyaudio
# On Windows: pip install pyaudio  (pre-built wheel usually works)
pyaudio>=0.2.14
